make mMSFPS functional; implement reconnection and resending settings on potential frame drops

+++ MiniScopeControlDlg.h
@@ -197,6 +197,6 @@ public:
 	afx_msg void OnEnKillfocusMinfluordisplay();
 	afx_msg void OnEnKillfocusMaxfluordisplay();
 	afx_msg void OnCbnCloseupCombo1();
-	int mMSFPS;
+	LONGLONG mMSFPS;
 	CComboBox mMSFPSCBox;
 };
